restructure occupancy_component according to home assistant guidelines
